requirements.txt
streamlit

transformers==4.57.0
torch
torchvision
accelerate
einops
sentencepiece

numpy==1.26.4
pillow==10.4.0
urllib3==1.26.18
python-dotenv